add test helpers
